graphics: Switch to 32-bit coordinates
On a 32-bit ARM, there is no benefit in computing on 16-bit, and
probably a net loss in some cases. Switch to 32-bit coordinates.
This also allows for better precision for some operations that
go "far in the woods".

Fixes: #334

+++ src/blitter.h
@@ -52,8 +52,8 @@ struct blitter
     // ========================================================================
 
     // Basic types
-    typedef int16_t  coord;
-    typedef uint16_t size;
+    typedef int32_t  coord;
+    typedef uint32_t size;
     typedef size_t   offset;
     typedef uint32_t pixword;
     typedef uint16_t palette_index;
